Hints for each move now being printed
Function in Board.hs was written to detect, for a given
colour and board, what moves result in a flip. These are added
to a list, which is returned. This function is called in Draw.hs
to check if a given tile is a valid move, and if so a different
bitmap is printed out